Checks if the projects list is empty before removing user permission when an OrgUnit is moved (fixes issue #567)

+++ sigmah/src/main/java/org/sigmah/server/policy/UserPermissionPolicy.java
@@ -144,7 +144,9 @@ public class UserPermissionPolicy {
 	public void deleteUserPermssionByOrgUnit(OrgUnit orgUnit){	 	
 		UserPermissionDAO permissionDAO = injector.getInstance(UserPermissionDAO.class);
 		List<Project> projects =  permissionDAO.getOrgUnitProjects(orgUnit);
-		permissionDAO.deleteByProjects(projects);		
+		if(!projects.isEmpty()) {
+			permissionDAO.deleteByProjects(projects);		
+		}
 	}
 	
 	/*
